[feat] #76 - 클아 구조, 전체 파일 정리
