+++ POTI-iOS/POTI-iOS/Domain/Entity/Order/OrderEntity.swift
@@ -5,18 +5,6 @@
 //  Created by mandoo on 1/20/26.
 //
 
-struct MemberEntity {
-    let id: Int
-    let name: String
-    let price: Int
-}
-
-struct ShippingEntity {
-    let id: Int
-    let name: String
-    let price: Int
-}
-
 struct OrderEntity {
     let members: [MemberEntity]
     let shippings: [ShippingEntity]
