+++ POTI-iOS/POTI-iOS/Domain/Interface/PostsInterface.swift
@@ -7,4 +7,6 @@
 
 protocol PostsInterface {
     func fetchOrderOptions(postId: Int) async throws -> PotOptionsEntity
+    
+    func fetchManagerData(postId: Int) async throws -> ManageEntity
 }
